--- a/server/server.yaml
+++ b/server/server.yaml
@@ -289,7 +289,8 @@ querier:
       enabled: true
       cache-item-size: 512000 # max size of cache item, unit: byte
       cache-max-count: 1024 # max capacity of cache list
-      cache-max-allow-deviation: 3600 # unit:s 
+      cache-first-timeout: 10 # time out for first cache item load, uint: s
+      cache-clean-interval: 3600 # clean interval for cache, unit: s
 
   # external-apm:
   # - name: skywalking