<!--

Thank you for contributing to DeepFlow!
Please read this template before submitting pull requests.
Texts surrounded by `<` and `>` should be replaced accordingly.
Put an `x` in `[ ]` to mark the item as checked. `[x]`

-->

### This PR is for:
- Server
<!--
One or more of:
- Agent
- CLI
- Server
- Message
- Libs
- Documents
- Workflow
-->

### Improves the performance of  promql query
- modify: 
- when first load cache, use `CacheKeyFoundNil` to mark cache loading, and return result instead of query when timeout
- when first load cache complete, use `channel` to trigger all waiters to continue, cancel `10ms` loops
- use `OR` instead of `REGEXP` when use `a|b|c` filters with regex
- optimize cache merge , use `map` to compare instead of `sort` first then write string.
- optimize `json unmarshal`, cache every json unmarshal result for unmarshalling
#### Added benchmark
- 
#### Benchmark result
query range: `1h`, step: `10m`
promql: 
```text
((sum by (cluster) (rate(apiserver_request_duration_seconds_count{job="apiserver",verb=~"LIST|GET"}[1h])) - ((sum by (cluster) (rate(apiserver_request_duration_seconds_bucket{job="apiserver",le="1",scope=~"resource|",verb=~"LIST|GET"}[1h])) or vector(0)) + sum by (cluster) (rate(apiserver_request_duration_seconds_bucket{job="apiserver",le="5",scope="namespace",verb=~"LIST|GET"}[1h])) + sum by (cluster) (rate(apiserver_request_duration_seconds_bucket{job="apiserver",le="30",scope="cluster",verb=~"LIST|GET"}[1h])))) + sum by (cluster) (rate(apiserver_request_total{code=~"5..",job="apiserver",verb=~"LIST|GET"}[1h]))) / sum by (cluster) (rate(apiserver_request_total{job="apiserver",verb=~"LIST|GET"}[1h]))
```

1. use cache ( mock 20 users, ~20qps）

- before modify:

|         | avg(ms) | max(ms) | min(ms) | p90 |
|---------|---------|---------|---------|-----|
| range   | 335     | 16941   | 23      | 514 |
| instant | 173     | 1289    | 17      | 370 |

![before](https://github.com/deepflowio/deepflow/assets/45836837/a2441655-b803-4c03-ac48-a45733a2a9ca)

- after modify: 

|         | avg(ms) | max(ms) | min(ms) | p90 |
|---------|---------|---------|---------|-----|
| range   | 217     | 10755   | 20      | 369 |
| instant | 116     | 1118    | 14      | 262 |

![after](https://github.com/deepflowio/deepflow/assets/45836837/1510b6e0-5d0d-49ca-a6fe-18667e4cd0c6)

2. not using cache ( mock 5 users, ~ 5qps)

- before:

|         | avg(ms) | max(ms) | min(ms) | p90   |
|---------|---------|---------|---------|-------|
| range   | 28351   | 35704   | 21233   | 32141 |
| instant | 16563   | 21775   | 13311   | 18798 |

![before-nocache-5](https://github.com/deepflowio/deepflow/assets/45836837/14f3be2c-d4ef-4eec-b59a-967275c00086)

- after:

|         | avg(ms) | max(ms) | min(ms) | p90   |
|---------|---------|---------|---------|-------|
| range   | 15682   | 21883   | 11390   | 19262 |
| instant | 10143   | 14659   | 7014    | 12402 |

![after-nocache-5](https://github.com/deepflowio/deepflow/assets/45836837/55e1568c-bc91-4515-a099-f29c569233a3)

3. Conclusion:
```text
avg lower 55% ~ 60% query time both in range & instant query
```